requirements.txt
fastapi 
uvicorn 
sqlalchemy 
mysql-connector

